Make managed interval util a hook
