+++ src/components/Clock.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { Divider, Tooltip } from 'antd'
 import { cyan, green, volcano } from '@ant-design/colors'
+import { useManagedInterval } from '../hooks/use-managed-interval'
 
 const eveTimeFormatter = new Intl.DateTimeFormat(
     undefined,
@@ -18,29 +19,6 @@ const makeCurrentTimes = () => {
         LOCAL: localTimeFormatter.format(now),
     };
 }
-const ONE_MINUTE = 60000;
-
-/**
- * Creates a managed setInterval which can be enabled or disabled at will.
- */
-const makeManagedInterval = (fn: Function) => {
-    let interval = null;
-
-    return {
-        enable: () => {
-            if (!interval) {
-                fn();
-                interval = setInterval(() => { fn(); }, ONE_MINUTE)
-            }
-        },
-        disable: () => {
-            if (interval) {
-                clearInterval(interval);
-                interval = null;
-            }
-        }
-    };
-};
 
 /**
  * For use with @ant-design/colors arrays.
@@ -59,25 +37,8 @@ const colorIndices = {
 const Clock = () => {
     const [currentTimes, setCurrentTimes] = useState(makeCurrentTimes);
 
-    useEffect(() => {
-        const clock = makeManagedInterval(() => setCurrentTimes(makeCurrentTimes()));
-
-        const ac = new AbortController();
-        document.addEventListener('visibilitychange', () => {
-            if (document.visibilityState === 'visible') {
-                clock.enable();
-            } else {
-                clock.disable()
-            }
-        }, { signal: ac.signal });
-
-        clock.enable();
-
-        return () => {
-            clock.disable()
-            ac.abort();
-        }
-    }, []);
+    const updateCurrentTimes = useCallback(() => setCurrentTimes(makeCurrentTimes()), []);
+    useManagedInterval(updateCurrentTimes)
 
     const borderColor = volcano.at(colorIndices.border);
     return (
